requirements.txt
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
expiringdict==1.1.4
fire==0.1.3
idna==2.8
monotonic==1.5
python-baseconv==1.2.1
python-icq-bot==0.0.19
requests==2.21.0
six==1.12.0
urllib3==1.24.1
